Bootloader: General Cleanup
