+++ .idea/QuantumOS.iml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="CPP_MODULE" version="4">
+  <component name="NewModuleRootManager">
+    <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/bootloader/src/bios_boot/stage-1/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/bootloader/src/bios_boot/stage-2/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/kernel/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
+      <excludeFolder url="file://$MODULE_DIR$/bootloader/src/bios_boot/stage-1/target" />
+      <excludeFolder url="file://$MODULE_DIR$/bootloader/src/bios_boot/stage-2/target" />
+      <excludeFolder url="file://$MODULE_DIR$/kernel/target" />
+      <excludeFolder url="file://$MODULE_DIR$/target" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+  </component>
+</module>
