+++ bootloader/src/bios_boot/stage-2/src/lib.rs
@@ -28,3 +28,4 @@ OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWA
 #![no_main] // disable all Rust-level entry points
 
 pub mod debug;
+pub mod paging;
