+++ bootloader/src/bios_boot/stage-2/src/main.rs
@@ -28,25 +28,17 @@ OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWA
 #![no_main] // disable all Rust-level entry points
 #![allow(dead_code)]
 
-use bootloader::boot_info::{BootInfo, VideoInformation};
-use core::arch::asm;
-use core::mem::size_of;
 use core::panic::PanicInfo;
-use lazy_static::lazy_static;
-use quantum_lib::address_utils::virtual_address::VirtAddress;
-use quantum_lib::bytes::Bytes;
+use bootloader::boot_info::{BootInfo, VideoInformation};
 use quantum_lib::debug::add_connection_to_global_stream;
-use quantum_lib::debug::stream_connection::{StreamConnection, StreamConnectionBuilder};
-use quantum_lib::x86_64::bios_call::BiosCall;
-use quantum_lib::x86_64::interrupts::Interrupts;
-use quantum_lib::x86_64::paging::config::PageConfigBuilder;
-use quantum_lib::x86_64::paging::structures::{PageMapLevel2, PageMapLevel3, PageMapLevel4};
-use quantum_lib::x86_64::registers::{SegmentRegs, CR0, CR3, CR4, EFLAGS, IA32_EFER};
+use quantum_lib::debug::stream_connection::StreamConnectionBuilder;
+use quantum_lib::debug_println;
+use quantum_lib::bytes::Bytes;
 use quantum_lib::x86_64::CPU;
-use quantum_lib::{debug_print, debug_println};
-use stage_2::debug::{display_string, setup_framebuffer};
 
 use spin::Mutex;
+use stage_2::debug::{display_string, setup_framebuffer};
+use stage_2::paging::enable_paging;
 
 #[no_mangle]
 #[link_section = ".start"]
@@ -112,102 +104,6 @@ fn main(boot_info: &BootInfo) {
     };
 
     debug_println!("Kernel ELF '{}'", test);
-}
-
-static mut LEVEL4: PageMapLevel4 = PageMapLevel4::new();
-static mut LEVEL3: PageMapLevel3 = PageMapLevel3::new();
-static mut LEVEL2: [PageMapLevel2; 5] = [PageMapLevel2::new(); 5];
-
-unsafe fn enable_paging() {
-    debug_print!("building pages ...");
-
-    let mut level4 = &mut LEVEL4;
-    let mut level3 = &mut LEVEL3;
-    let mut level2_tables = &mut LEVEL2;
-
-    for (offset, level2) in level2_tables.iter_mut().enumerate() {
-        let offset_addition = offset as u64 * 1024 * 1024 * 1024;
-
-        for i in 0..512 {
-            let huge_address = VirtAddress::new(i * 2 * 1024 * 1024 + offset_addition)
-                .unwrap()
-                .try_aligned()
-                .unwrap();
-
-            let two_mb_entries = PageConfigBuilder::new()
-                .level2()
-                .present(true)
-                .read_write(true)
-                .executable(true)
-                .user_page(false)
-                .set_huge_page_address(huge_address)
-                .build()
-                .unwrap();
-
-            level2.set_entry(two_mb_entries, i as usize).unwrap();
-        }
-
-        debug_print!("L2...");
-
-        let level_2_entry = PageConfigBuilder::new()
-            .level3()
-            .present(true)
-            .read_write(true)
-            .executable(true)
-            .user_page(false)
-            .set_address_of_next_table(level2.get_address())
-            .build()
-            .unwrap();
-
-        level3.set_entry(level_2_entry, offset).unwrap();
-    }
-
-    debug_print!("L3...");
-
-    let level_3_config = PageConfigBuilder::new()
-        .level4()
-        .present(true)
-        .read_write(true)
-        .executable(true)
-        .user_page(false)
-        .set_address_of_next_table(level3.get_address())
-        .build()
-        .unwrap();
-
-    level4.set_entry(level_3_config, 0).unwrap();
-
-    debug_print!("L4...");
-    let level4_address = level4.get_address().as_u64();
-    debug_println!(" OK ({}Gib Mapped!)", level2_tables.len());
-
-    debug_print!("Disabling paging ...");
-    CR0::set_paging(false);
-    debug_println!("OK");
-
-    debug_print!("Setting PAE ...");
-    CR4::set_physical_address_extension(true);
-    CR4::set_page_size_extension(true);
-    debug_println!("OK");
-
-    debug_print!("Setting Long mode ...");
-    IA32_EFER::set_long_mode_enable(true);
-    debug_println!("OK");
-
-    debug_print!("Loading CR3 ...");
-    CR3::set_page_directory_base_register(level4_address as *mut u8);
-    debug_println!("OK 0x{:x}", level4_address);
-
-    debug_print!("Enabling protected mode ...");
-    CR0::set_protected_mode(true);
-    debug_println!("OK");
-
-    debug_print!("Enabling paging ...");
-    CR0::set_paging(true);
-    debug_println!("OK");
-
-    debug_print!("Reloading segment registers ...");
-    SegmentRegs::reload_all_to(0x10);
-    debug_println!("OK");
 }
 
 #[panic_handler]
